+++ src/ShapeNetwork/heap.cpp
@@ -2,14 +2,14 @@
 # include "heap.h"
 
 heap::heap()
-{	
+{
 }
 
 heap::~heap()
 {
 }
 void heap::clear()
-{	
+{
 	all_nodes.clear();
 }
 
@@ -22,19 +22,20 @@ void heap::insert( long triangleIndex, double value )
 }
 
 long heap::size()
-{	return all_nodes.size();
+{
+	return static_cast<long>(all_nodes.size());
 }
 
 inline void heap::trickleUp()
-{	
-	long index = all_nodes.size() - 1 ;
+{
+	long index = static_cast<long>(all_nodes.size()) - 1 ;
 	long parent;
 
 	bool trickle = true;
-	heapnode tempNode;	
+	heapnode tempNode;
 
 	while( trickle )
-	{		
+	{
 		//Even Kid
 		if( index % 2 == 0 )
 			parent = ( index - 2 ) / 2;//*.5;
@@ -58,7 +59,8 @@ inline void heap::trickleUp()
 }
 
 void heap::pop()
-{	all_nodes.pop_front();
+{
+	all_nodes.pop_front();
 	if( all_nodes.size() > 1 )
 	{	all_nodes.push_front( all_nodes[ all_nodes.size() - 1 ] );
 		all_nodes.pop_back();
@@ -67,14 +69,15 @@ void heap::pop()
 }
 
 heapnode heap::top()
-{	if( all_nodes.size() > 0 )
+{
+	if( all_nodes.size() > 0 )
 		return all_nodes[0];
 	else
 		return heapnode();
 }
 
 inline void heap::trickleDown()
-{	
+{
 	long index = 0;
 	bool trickle = true;
 	long swap_index;
@@ -85,7 +88,7 @@ inline void heap::trickleDown()
 		long kid2 = kid1 + 1;
 
 		swap_index = minValIndex( kid1, kid2 );
-		if( swap_index > 0 && swap_index < (int)all_nodes.size() )
+		if( swap_index > 0 && swap_index < static_cast<int>(all_nodes.size()) )
 		{
 			if( all_nodes[index] > all_nodes[swap_index] )
 			{	tempNode = all_nodes[index];
@@ -103,7 +106,7 @@ inline void heap::trickleDown()
 
 inline long heap::minValIndex( long kid1, long kid2 )
 {
-	long heap_size = all_nodes.size();
+	long heap_size = static_cast<long>(all_nodes.size());
 	if( kid2 >= heap_size || kid1 >= heap_size )
 		return kid1;	
 	else
