+++ src/Tin/TinHeap.cpp
@@ -11,7 +11,7 @@ TinHeap::~TinHeap()
 {
 }
 void TinHeap::clear()
-{	
+{
 	all_nodes.clear();
 }
 
@@ -27,15 +27,15 @@ void TinHeap::insert( long triangleIndex, double dev, vertex v )
 }
 
 inline void TinHeap::trickleUp()
-{	
-	long index = all_nodes.size() - 1 ;
+{
+	long index = static_cast<long>(all_nodes.size()) - 1 ;
 	long parent;
 
 	bool trickle = true;
 	heapNode tempNode;	
 
 	while( trickle )
-	{		
+	{
 		//Even Kid
 		if( index % 2 == 0 )
 			parent = ( index - 2 ) / 2; //*.5;
@@ -65,7 +65,7 @@ void TinHeap::pop()
 	{	all_nodes.push_front( all_nodes[ all_nodes.size() - 1 ] );
 		all_nodes.pop_back();
 		trickleDown();
-	}	
+	}
 }
 
 heapNode TinHeap::top()
@@ -87,12 +87,12 @@ inline void TinHeap::trickleDown()
 		long kid2 = kid1 + 1;
 
 		swap_index = maxValIndex( kid1, kid2 );
-		if( swap_index > 0 && swap_index < (int)all_nodes.size() )
+		if( swap_index > 0 && swap_index < static_cast<int>(all_nodes.size()) )
 		{
 			if( all_nodes[index] < all_nodes[swap_index] )
 			{	tempNode = all_nodes[index];
 				all_nodes[index] = all_nodes[swap_index];
-				all_nodes[swap_index] = tempNode;			
+				all_nodes[swap_index] = tempNode;
 				index = swap_index;
 			}
 			else
@@ -105,11 +105,12 @@ inline void TinHeap::trickleDown()
 
 inline long TinHeap::maxValIndex( long kid1, long kid2 )
 {
-	long heap_size = all_nodes.size();
+	long heap_size = static_cast<long>(all_nodes.size());
 	if( kid2 >= heap_size || kid1 >= heap_size )
 		return kid1;	
 	else
-	{	if( all_nodes[kid1] > all_nodes[kid2] )
+	{
+		if( all_nodes[kid1] > all_nodes[kid2] )
 			return kid1;
 		else
 			return kid2;
